rename viewModel,fix psbtFragment bugs,optimize structure
