+++ app/src/main/java/com/keystone/cold/viewmodel/ParsePsbtViewModel.java
@@ -1,8 +1,8 @@
 package com.keystone.cold.viewmodel;
 
 import android.app.Application;
+import android.os.Bundle;
 import android.text.TextUtils;
-import android.util.Log;
 
 import androidx.annotation.NonNull;
 import androidx.lifecycle.AndroidViewModel;
@@ -10,32 +10,20 @@ import androidx.lifecycle.MutableLiveData;
 
 import com.googlecode.protobuf.format.JsonFormat;
 import com.keystone.coinlib.Util;
-import com.keystone.coinlib.coins.AbsCoin;
 import com.keystone.coinlib.coins.AbsTx;
-import com.keystone.coinlib.coins.BTC.Btc;
-import com.keystone.coinlib.coins.BTC.BtcImpl;
-import com.keystone.coinlib.coins.BTC.Deriver;
 import com.keystone.coinlib.coins.BTC.UtxoTx;
-import com.keystone.coinlib.exception.InvalidPathException;
-import com.keystone.coinlib.interfaces.SignCallback;
 import com.keystone.coinlib.interfaces.Signer;
-import com.keystone.coinlib.path.AddressIndex;
-import com.keystone.coinlib.path.CoinPath;
 import com.keystone.coinlib.utils.Account;
-import com.keystone.coinlib.utils.Coins;
 import com.keystone.cold.AppExecutors;
 import com.keystone.cold.DataRepository;
 import com.keystone.cold.MainApplication;
 import com.keystone.cold.Utilities;
-import com.keystone.cold.callables.ClearTokenCallable;
 import com.keystone.cold.callables.GetMessageCallable;
 import com.keystone.cold.callables.GetPasswordTokenCallable;
 import com.keystone.cold.callables.VerifyFingerprintCallable;
 import com.keystone.cold.db.entity.AccountEntity;
-import com.keystone.cold.db.entity.AddressEntity;
 import com.keystone.cold.db.entity.CoinEntity;
 import com.keystone.cold.db.entity.TxEntity;
-import com.keystone.cold.encryption.ChipSigner;
 import com.keystone.cold.protobuf.TransactionProtoc;
 import com.keystone.cold.ui.views.AuthenticateModal;
 
@@ -45,21 +33,16 @@ import org.json.JSONObject;
 import org.spongycastle.util.encoders.Hex;
 
 import java.security.SignatureException;
-import java.text.NumberFormat;
 import java.util.Comparator;
 import java.util.List;
-import java.util.Objects;
 import java.util.Optional;
-import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
-import java.util.stream.Stream;
 
 import static com.keystone.cold.ui.fragment.main.FeeAttackChecking.FeeAttackCheckingResult.DUPLICATE_TX;
 import static com.keystone.cold.ui.fragment.main.FeeAttackChecking.FeeAttackCheckingResult.NORMAL;
 import static com.keystone.cold.ui.fragment.main.FeeAttackChecking.FeeAttackCheckingResult.SAME_OUTPUTS;
-import static com.keystone.cold.viewmodel.AddAddressViewModel.AddAddressTask.getAddressType;
 import static com.keystone.cold.viewmodel.GlobalViewModel.getAccount;
 import static com.keystone.cold.viewmodel.WatchWallet.ELECTRUM;
 
@@ -70,7 +53,6 @@ public abstract class ParsePsbtViewModel extends AndroidViewModel {
     public static final String STATE_SIGN_FAIL = "signing_fail";
     public static final String STATE_SIGN_SUCCESS = "signing_success";
     protected final DataRepository mRepository;
-    protected final MutableLiveData<TxEntity> observableTx = new MutableLiveData<>();
     protected final MutableLiveData<Integer> feeAttachCheckingResult = new MutableLiveData<>();
     protected final MutableLiveData<Boolean> addingAddress = new MutableLiveData<>();
     protected final MutableLiveData<String> signState = new MutableLiveData<>();
@@ -83,11 +65,6 @@ public abstract class ParsePsbtViewModel extends AndroidViewModel {
     public ParsePsbtViewModel(@NonNull Application application) {
         super(application);
         mRepository = MainApplication.getApplication().getRepository();
-        observableTx.setValue(null);
-    }
-
-    public MutableLiveData<TxEntity> getObservableTx() {
-        return observableTx;
     }
 
     public MutableLiveData<Integer> getFeeAttachCheckingResult() {
@@ -121,147 +98,6 @@ public abstract class ParsePsbtViewModel extends AndroidViewModel {
         });
     }
 
-    protected TxEntity generateTxEntity(JSONObject object) throws JSONException {
-        TxEntity tx = new TxEntity();
-        NumberFormat nf = NumberFormat.getInstance();
-        nf.setMaximumFractionDigits(20);
-        coinCode = Objects.requireNonNull(transaction).getCoinCode();
-        tx.setSignId(object.getString("signId"));
-        tx.setTimeStamp(object.optLong("timestamp"));
-        tx.setCoinCode(coinCode);
-        tx.setCoinId(Coins.coinIdFromCoinCode(coinCode));
-        tx.setFrom(getFromAddress());
-        tx.setTo(getToAddress());
-        tx.setAmount(nf.format(transaction.getAmount()) + " " + transaction.getUnit());
-        tx.setFee(nf.format(transaction.getFee()) + " BTC");
-        tx.setMemo(transaction.getMemo());
-        tx.setBelongTo(mRepository.getBelongTo());
-        return tx;
-    }
-
-    protected boolean isExternalPath(@NonNull String path) {
-        try {
-            return CoinPath.parsePath(path).getParent().isExternal();
-        } catch (InvalidPathException e) {
-            e.printStackTrace();
-        }
-        return false;
-    }
-
-    protected String getFromAddress() {
-        if (!TextUtils.isEmpty(transaction.getFrom())) {
-            return transaction.getFrom();
-        }
-        String[] paths = transaction.getHdPath().split(AbsTx.SEPARATOR);
-        String[] externalPath = Stream.of(paths)
-                .filter(this::isExternalPath)
-                .toArray(String[]::new);
-        ensureAddressExist(externalPath);
-
-        try {
-            if (transaction instanceof UtxoTx) {
-                JSONArray inputsClone = new JSONArray();
-                JSONArray inputs = ((UtxoTx) transaction).getInputs();
-
-                CoinEntity coinEntity = mRepository.loadCoinSync(Coins.coinIdFromCoinCode(transaction.getCoinCode()));
-                AccountEntity accountEntity =
-                        mRepository.loadAccountsByPath(coinEntity.getId(), getAccount(getApplication()).getPath());
-
-                for (int i = 0; i < inputs.length(); i++) {
-                    JSONObject input = inputs.getJSONObject(i);
-                    long value = input.getJSONObject("utxo").getLong("value");
-                    String hdpath = input.getString("ownerKeyPath");
-                    AddressIndex addressIndex = CoinPath.parsePath(hdpath);
-                    int index = addressIndex.getValue();
-                    int change = addressIndex.getParent().getValue();
-
-                    String from = new Deriver(Utilities.isMainNet(getApplication())).derive(accountEntity.getExPub()
-                            , change, index, getAddressType(accountEntity));
-                    inputsClone.put(new JSONObject().put("value", value)
-                            .put("address", from));
-                }
-
-                return inputsClone.toString();
-            }
-        } catch (JSONException | InvalidPathException e) {
-            e.printStackTrace();
-        }
-
-        return Stream.of(externalPath)
-                .distinct()
-                .map(path -> mRepository.loadAddressBypath(path).getAddressString())
-                .reduce((s1, s2) -> s1 + AbsTx.SEPARATOR + s2)
-                .orElse("");
-    }
-
-    protected void ensureAddressExist(String[] paths) {
-        if (paths == null || paths.length == 0) {
-            return;
-        }
-        String maxIndexHdPath = paths[0];
-        if (paths.length > 1) {
-            int max = getAddressIndex(paths[0]);
-            for (String path : paths) {
-                if (getAddressIndex(path) > max) {
-                    max = getAddressIndex(path);
-                    maxIndexHdPath = path;
-                }
-            }
-        }
-        AddressEntity address = mRepository.loadAddressBypath(maxIndexHdPath);
-        if (address == null) {
-            addAddress(maxIndexHdPath);
-        }
-    }
-
-    protected void addAddress(String hdPath) {
-        String accountHdPath;
-        int pathIndex;
-        try {
-            AddressIndex index = CoinPath.parsePath(hdPath);
-            pathIndex = index.getValue();
-            accountHdPath = index.getParent().getParent().toString();
-        } catch (InvalidPathException e) {
-            e.printStackTrace();
-            return;
-        }
-
-        CoinEntity coin = mRepository.loadCoinSync(Coins.coinIdFromCoinCode(coinCode));
-        AccountEntity accountEntity = getAccountEntityByPath(accountHdPath, coin);
-        if (accountEntity == null) return;
-
-        List<AddressEntity> addressEntities = mRepository.loadAddressSync(coin.getCoinId());
-        Optional<AddressEntity> addressEntityOptional = addressEntities
-                .stream()
-                .filter(addressEntity -> addressEntity.getPath()
-                        .startsWith(accountEntity.getHdPath() + "/" + 0))
-                .max((o1, o2) -> o1.getPath().compareTo(o2.getPath()));
-
-        int index = -1;
-        if (addressEntityOptional.isPresent()) {
-            try {
-                AddressIndex addressIndex = CoinPath.parsePath(addressEntityOptional.get().getPath());
-                index = addressIndex.getValue();
-            } catch (InvalidPathException e) {
-                e.printStackTrace();
-            }
-        }
-
-        if (index < pathIndex) {
-            final CountDownLatch mLatch = new CountDownLatch(1);
-            addingAddress.postValue(true);
-            new AddAddressViewModel.AddAddressTask(coin, mRepository, mLatch::countDown,
-                    accountEntity.getExPub(), 0).execute(pathIndex - index);
-            try {
-                mLatch.await();
-                addingAddress.postValue(false);
-            } catch (InterruptedException e) {
-                e.printStackTrace();
-            }
-        }
-
-    }
-
     protected int getAddressIndex(String hdPath) {
         String[] splits = hdPath.split("/");
         try {
@@ -273,7 +109,6 @@ public abstract class ParsePsbtViewModel extends AndroidViewModel {
         return 0;
     }
 
-
     protected String getToAddress() {
         String to = transaction.getTo();
         if (transaction instanceof UtxoTx) {
@@ -283,68 +118,6 @@ public abstract class ParsePsbtViewModel extends AndroidViewModel {
             }
         }
         return to;
-    }
-
-    protected boolean checkChangeAddress(AbsTx utxoTx) {
-        List<UtxoTx.ChangeAddressInfo> changeAddressInfoList = ((UtxoTx) utxoTx).getChangeAddressInfo();
-
-        if (changeAddressInfoList == null || changeAddressInfoList.isEmpty()) {
-            return true;
-        }
-        Deriver deriver = new Deriver(Utilities.isMainNet(getApplication()));
-        for (UtxoTx.ChangeAddressInfo changeAddressInfo : changeAddressInfoList) {
-            String hdPath = changeAddressInfo.hdPath;
-            String address = changeAddressInfo.address;
-            String accountHdPath = getAccountHdPath(changeAddressInfo.hdPath);
-            AccountEntity accountEntity = getAccountEntityByPath(accountHdPath,
-                    mRepository.loadCoinEntityByCoinCode(utxoTx.getCoinCode()));
-            if (accountEntity == null) {
-                return false;
-            }
-            String exPub = accountEntity.getExPub();
-            try {
-                AddressIndex addressIndex = CoinPath.parsePath(hdPath);
-                int change = addressIndex.getParent().getValue();
-                int index = addressIndex.getValue();
-                String expectAddress = Objects.requireNonNull(deriver).derive(exPub, change,
-                        index, getAddressType(accountEntity));
-                if (!address.equals(expectAddress)) {
-                    return false;
-                }
-            } catch (InvalidPathException e) {
-                e.printStackTrace();
-                return false;
-            }
-        }
-
-        return true;
-    }
-
-    protected AccountEntity getAccountEntityByPath(String accountHdPath, CoinEntity coin) {
-        List<AccountEntity> accountEntities = mRepository.loadAccountsForCoin(coin);
-        Optional<AccountEntity> optional = accountEntities.stream()
-                .filter(accountEntity ->
-                        accountEntity.getHdPath().equals(accountHdPath.toUpperCase()))
-                .findFirst();
-
-        AccountEntity accountEntity;
-        if (optional.isPresent()) {
-            accountEntity = optional.get();
-        } else {
-            return null;
-        }
-        return accountEntity;
-    }
-
-    protected String getAccountHdPath(String addressPath) {
-        String accountHdPath;
-        try {
-            accountHdPath = CoinPath.parsePath(addressPath).getParent().getParent().toString();
-        } catch (InvalidPathException e) {
-            e.printStackTrace();
-            return null;
-        }
-        return accountHdPath;
     }
 
     protected JSONObject parsePsbtTx(JSONObject adaptTx) throws JSONException {
@@ -367,7 +140,7 @@ public abstract class ParsePsbtViewModel extends AndroidViewModel {
         return signTx;
     }
 
-    protected long generateAutoIncreaseId() {
+    private long generateAutoIncreaseId() {
         List<TxEntity> txEntityList = mRepository.loadAllTxsSync(Utilities.currentCoin(getApplication()).coinId());
         if (txEntityList == null || txEntityList.isEmpty()) {
             return 0;
@@ -389,45 +162,6 @@ public abstract class ParsePsbtViewModel extends AndroidViewModel {
 
     public void setToken(AuthenticateModal.OnVerify.VerifyToken token) {
         this.token = token;
-    }
-
-    public void handleSign() {
-        AppExecutors.getInstance().diskIO().execute(() -> {
-            SignCallback callback = initSignCallback();
-            callback.startSign();
-            Signer[] signer = initSigners();
-            signTransaction(transaction, callback, signer);
-        });
-    }
-
-    protected SignCallback initSignCallback() {
-        return new SignCallback() {
-            @Override
-            public void startSign() {
-                signState.postValue(STATE_SIGNING);
-            }
-
-            @Override
-            public void onFail() {
-                signState.postValue(STATE_SIGN_FAIL);
-                new ClearTokenCallable().call();
-            }
-
-            @Override
-            public void onSuccess(String txId, String rawTx) {
-                TxEntity tx = observableTx.getValue();
-                Objects.requireNonNull(tx).setTxId(txId);
-                tx.setSignedHex(rawTx);
-                mRepository.insertTx(tx);
-                signState.postValue(STATE_SIGN_SUCCESS);
-                new ClearTokenCallable().call();
-            }
-
-            @Override
-            public void postProgress(int progress) {
-
-            }
-        };
     }
 
     protected String getAuthToken() {
@@ -453,24 +187,25 @@ public abstract class ParsePsbtViewModel extends AndroidViewModel {
         return authToken;
     }
 
-    private void signTransaction(@NonNull AbsTx transaction, @NonNull SignCallback callback, Signer... signer) {
-        if (signer == null) {
-            callback.onFail();
-            return;
+    protected AccountEntity getAccountEntityByPath(String accountHdPath, CoinEntity coin) {
+        List<AccountEntity> accountEntities = mRepository.loadAccountsForCoin(coin);
+        Optional<AccountEntity> optional = accountEntities.stream()
+                .filter(accountEntity ->
+                        accountEntity.getHdPath().equals(accountHdPath.toUpperCase()))
+                .findFirst();
+
+        AccountEntity accountEntity;
+        if (optional.isPresent()) {
+            accountEntity = optional.get();
+        } else {
+            return null;
         }
-        switch (transaction.getTxType()) {
-            case "OMNI":
-            case "OMNI_USDT":
-                Btc btc = new Btc(new BtcImpl(Utilities.isMainNet(getApplication())));
-                btc.generateOmniTx(transaction, callback, signer);
-                break;
-            default:
-                AbsCoin coin = AbsCoin.newInstance(coinCode);
-                Objects.requireNonNull(coin).generateTransaction(transaction, callback, signer);
-        }
+        return accountEntity;
     }
+
+    public abstract void parseTxData(Bundle bundle);
 
     public abstract void handleSignPsbt(String psbt);
 
-    protected abstract Signer[] initSigners();
+    public abstract Signer[] initSigners();
 }
