+++ app/src/main/java/com/keystone/cold/viewmodel/PsbtSigleConfirmViewModel.java
@@ -1,0 +1,376 @@
+package com.keystone.cold.viewmodel;
+
+import android.app.Application;
+import android.os.Bundle;
+import android.text.TextUtils;
+import android.util.Log;
+
+import androidx.annotation.NonNull;
+import androidx.lifecycle.MutableLiveData;
+
+import com.keystone.coinlib.Util;
+import com.keystone.coinlib.coins.AbsTx;
+import com.keystone.coinlib.coins.BTC.Btc;
+import com.keystone.coinlib.coins.BTC.BtcImpl;
+import com.keystone.coinlib.coins.BTC.Deriver;
+import com.keystone.coinlib.coins.BTC.UtxoTx;
+import com.keystone.coinlib.exception.InvalidPathException;
+import com.keystone.coinlib.exception.InvalidTransactionException;
+import com.keystone.coinlib.interfaces.SignPsbtCallback;
+import com.keystone.coinlib.interfaces.Signer;
+import com.keystone.coinlib.path.AddressIndex;
+import com.keystone.coinlib.path.CoinPath;
+import com.keystone.coinlib.utils.Coins;
+import com.keystone.cold.AppExecutors;
+import com.keystone.cold.Utilities;
+import com.keystone.cold.callables.ClearTokenCallable;
+import com.keystone.cold.db.entity.AccountEntity;
+import com.keystone.cold.db.entity.AddressEntity;
+import com.keystone.cold.db.entity.CoinEntity;
+import com.keystone.cold.db.entity.TxEntity;
+import com.keystone.cold.encryption.ChipSigner;
+import com.keystone.cold.ui.fragment.main.adapter.PsbtTxAdapter;
+import com.keystone.cold.viewmodel.exceptions.WatchWalletNotMatchException;
+
+import org.json.JSONArray;
+import org.json.JSONException;
+import org.json.JSONObject;
+
+import java.text.NumberFormat;
+import java.util.List;
+import java.util.Objects;
+import java.util.Optional;
+import java.util.concurrent.CountDownLatch;
+import java.util.stream.Stream;
+
+import static com.keystone.cold.viewmodel.AddAddressViewModel.AddAddressTask.getAddressType;
+import static com.keystone.cold.viewmodel.GlobalViewModel.getAccount;
+import static com.keystone.cold.viewmodel.WatchWallet.ELECTRUM;
+
+public class PsbtSigleConfirmViewModel extends ParsePsbtViewModel {
+    private static final String TAG = "SigleTxConfirmViewModel";
+    protected final MutableLiveData<TxEntity> observableTx = new MutableLiveData<>();
+
+    public PsbtSigleConfirmViewModel(@NonNull Application application) {
+        super(application);
+        observableTx.setValue(null);
+    }
+
+    @Override
+    public void parseTxData(Bundle bundle) {
+        AppExecutors.getInstance().diskIO().execute(() -> {
+            try {
+                String psbtBase64 = bundle.getString("psbt_base64");
+                Btc btc = new Btc(new BtcImpl(Utilities.isMainNet(getApplication())));
+                JSONObject psbtTx = btc.parsePsbt(psbtBase64);
+                if (psbtTx == null) {
+                    parseTxException.postValue(new InvalidTransactionException("parse failed,invalid psbt data"));
+                    return;
+                }
+                boolean isMultigisTx = psbtTx.getJSONArray("inputs").getJSONObject(0).getBoolean("isMultiSign");
+                if (isMultigisTx) {
+                    parseTxException.postValue(
+                            new InvalidTransactionException("", InvalidTransactionException.IS_MULTISIG_TX));
+                    return;
+                }
+                JSONObject adaptTx = new PsbtTxAdapter().adapt(psbtTx);
+                if (adaptTx.getJSONArray("inputs").length() == 0) {
+                    parseTxException.postValue(
+                            new InvalidTransactionException("master xfp not match, or nothing can be sign"));
+                    return;
+                }
+                JSONObject signTx = parsePsbtTx(adaptTx);
+                Log.i(TAG, "signTx = " + signTx.toString(4));
+                transaction = AbsTx.newInstance(signTx);
+                if (transaction == null) {
+                    observableTx.postValue(null);
+                    parseTxException.postValue(new InvalidTransactionException("invalid transaction"));
+                    return;
+                }
+                if (transaction instanceof UtxoTx) {
+                    if (!checkChangeAddress(transaction)) {
+                        observableTx.postValue(null);
+                        parseTxException.postValue(new InvalidTransactionException("invalid change address"));
+                        return;
+                    }
+                }
+                TxEntity tx = generateTxEntity(signTx);
+                observableTx.postValue(tx);
+                if (Coins.BTC.coinCode().equals(transaction.getCoinCode())
+                        || Coins.XTN.coinCode().equals(transaction.getCoinCode())) {
+                    feeAttackChecking(tx);
+                }
+            } catch (JSONException e) {
+                e.printStackTrace();
+                parseTxException.postValue(new InvalidTransactionException("adapt failed,invalid psbt data"));
+            } catch (WatchWalletNotMatchException e) {
+                e.printStackTrace();
+                parseTxException.postValue(e);
+            }
+        });
+    }
+
+    @Override
+    public void handleSignPsbt(String psbt) {
+        AppExecutors.getInstance().diskIO().execute(() -> {
+            SignPsbtCallback callback = new SignPsbtCallback() {
+                @Override
+                public void startSign() {
+                    signState.postValue(STATE_SIGNING);
+                }
+
+                @Override
+                public void onFail() {
+                    signState.postValue(STATE_SIGN_FAIL);
+                    new ClearTokenCallable().call();
+                }
+
+                @Override
+                public void onSuccess(String txId, String psbtB64) {
+                    TxEntity tx = observableTx.getValue();
+                    Objects.requireNonNull(tx);
+                    if (TextUtils.isEmpty(txId)) {
+                        txId = "unknown_txid_" + Math.abs(tx.hashCode());
+                    }
+                    tx.setTxId(txId);
+                    tx.setSignedHex(psbtB64);
+                    mRepository.insertTx(tx);
+                    signState.postValue(STATE_SIGN_SUCCESS);
+                    new ClearTokenCallable().call();
+                }
+
+                @Override
+                public void postProgress(int progress) {
+
+                }
+            };
+            callback.startSign();
+            Signer[] signer = initSigners();
+            Btc btc = new Btc(new BtcImpl(Utilities.isMainNet(getApplication())));
+            if (WatchWallet.getWatchWallet(getApplication()) == ELECTRUM) {
+                btc.signPsbt(psbt, callback, false, signer);
+            } else {
+                btc.signPsbt(psbt, callback, signer);
+            }
+        });
+    }
+
+    @Override
+    public Signer[] initSigners() {
+        String[] paths = transaction.getHdPath().split(AbsTx.SEPARATOR);
+        String coinCode = transaction.getCoinCode();
+        String[] distinctPaths = Stream.of(paths).distinct().toArray(String[]::new);
+        Signer[] signer = new Signer[distinctPaths.length];
+
+        String authToken = getAuthToken();
+        if (TextUtils.isEmpty(authToken)) {
+            Log.w(TAG, "authToken null");
+            return null;
+        }
+        CoinEntity coinEntity = mRepository.loadCoinEntityByCoinCode(coinCode);
+        for (int i = 0; i < distinctPaths.length; i++) {
+            String accountHdPath = getAccountHdPath(distinctPaths[i]);
+            if (accountHdPath == null) {
+                return null;
+            }
+            AccountEntity accountEntity = getAccountEntityByPath(accountHdPath, coinEntity);
+            if (accountEntity == null) {
+                return null;
+            }
+            String pubKey = Util.getPublicKeyHex(accountEntity.getExPub(), distinctPaths[i]);
+            signer[i] = new ChipSigner(distinctPaths[i].toLowerCase(), authToken, pubKey);
+        }
+        return signer;
+    }
+
+    private boolean checkChangeAddress(AbsTx utxoTx) {
+        List<UtxoTx.ChangeAddressInfo> changeAddressInfoList = ((UtxoTx) utxoTx).getChangeAddressInfo();
+
+        if (changeAddressInfoList == null || changeAddressInfoList.isEmpty()) {
+            return true;
+        }
+        Deriver deriver = new Deriver(Utilities.isMainNet(getApplication()));
+        for (UtxoTx.ChangeAddressInfo changeAddressInfo : changeAddressInfoList) {
+            String hdPath = changeAddressInfo.hdPath;
+            String address = changeAddressInfo.address;
+            String accountHdPath = getAccountHdPath(changeAddressInfo.hdPath);
+            AccountEntity accountEntity = getAccountEntityByPath(accountHdPath,
+                    mRepository.loadCoinEntityByCoinCode(utxoTx.getCoinCode()));
+            if (accountEntity == null) {
+                return false;
+            }
+            String exPub = accountEntity.getExPub();
+            try {
+                AddressIndex addressIndex = CoinPath.parsePath(hdPath);
+                int change = addressIndex.getParent().getValue();
+                int index = addressIndex.getValue();
+                String expectAddress = Objects.requireNonNull(deriver).derive(exPub, change,
+                        index, getAddressType(accountEntity));
+                if (!address.equals(expectAddress)) {
+                    return false;
+                }
+            } catch (InvalidPathException e) {
+                e.printStackTrace();
+                return false;
+            }
+        }
+
+        return true;
+    }
+
+    private String getAccountHdPath(String addressPath) {
+        String accountHdPath;
+        try {
+            accountHdPath = CoinPath.parsePath(addressPath).getParent().getParent().toString();
+        } catch (InvalidPathException e) {
+            e.printStackTrace();
+            return null;
+        }
+        return accountHdPath;
+    }
+
+    private TxEntity generateTxEntity(JSONObject object) throws JSONException {
+        TxEntity tx = new TxEntity();
+        NumberFormat nf = NumberFormat.getInstance();
+        nf.setMaximumFractionDigits(20);
+        coinCode = Objects.requireNonNull(transaction).getCoinCode();
+        tx.setSignId(object.getString("signId"));
+        tx.setTimeStamp(object.optLong("timestamp"));
+        tx.setCoinCode(coinCode);
+        tx.setCoinId(Coins.coinIdFromCoinCode(coinCode));
+        tx.setFrom(getFromAddress());
+        tx.setTo(getToAddress());
+        tx.setAmount(nf.format(transaction.getAmount()) + " " + transaction.getUnit());
+        tx.setFee(nf.format(transaction.getFee()) + " BTC");
+        tx.setMemo(transaction.getMemo());
+        tx.setBelongTo(mRepository.getBelongTo());
+        return tx;
+    }
+
+    private boolean isExternalPath(@NonNull String path) {
+        try {
+            return CoinPath.parsePath(path).getParent().isExternal();
+        } catch (InvalidPathException e) {
+            e.printStackTrace();
+        }
+        return false;
+    }
+
+    private String getFromAddress() {
+        if (!TextUtils.isEmpty(transaction.getFrom())) {
+            return transaction.getFrom();
+        }
+        String[] paths = transaction.getHdPath().split(AbsTx.SEPARATOR);
+        String[] externalPath = Stream.of(paths)
+                .filter(this::isExternalPath)
+                .toArray(String[]::new);
+        ensureAddressExist(externalPath);
+
+        try {
+            if (transaction instanceof UtxoTx) {
+                JSONArray inputsClone = new JSONArray();
+                JSONArray inputs = ((UtxoTx) transaction).getInputs();
+
+                CoinEntity coinEntity = mRepository.loadCoinSync(Coins.coinIdFromCoinCode(transaction.getCoinCode()));
+                AccountEntity accountEntity =
+                        mRepository.loadAccountsByPath(coinEntity.getId(), getAccount(getApplication()).getPath());
+
+                for (int i = 0; i < inputs.length(); i++) {
+                    JSONObject input = inputs.getJSONObject(i);
+                    long value = input.getJSONObject("utxo").getLong("value");
+                    String hdpath = input.getString("ownerKeyPath");
+                    AddressIndex addressIndex = CoinPath.parsePath(hdpath);
+                    int index = addressIndex.getValue();
+                    int change = addressIndex.getParent().getValue();
+
+                    String from = new Deriver(Utilities.isMainNet(getApplication())).derive(accountEntity.getExPub()
+                            , change, index, getAddressType(accountEntity));
+                    inputsClone.put(new JSONObject().put("value", value)
+                            .put("address", from));
+                }
+
+                return inputsClone.toString();
+            }
+        } catch (JSONException | InvalidPathException e) {
+            e.printStackTrace();
+        }
+
+        return Stream.of(externalPath)
+                .distinct()
+                .map(path -> mRepository.loadAddressBypath(path).getAddressString())
+                .reduce((s1, s2) -> s1 + AbsTx.SEPARATOR + s2)
+                .orElse("");
+    }
+
+    private void ensureAddressExist(String[] paths) {
+        if (paths == null || paths.length == 0) {
+            return;
+        }
+        String maxIndexHdPath = paths[0];
+        if (paths.length > 1) {
+            int max = getAddressIndex(paths[0]);
+            for (String path : paths) {
+                if (getAddressIndex(path) > max) {
+                    max = getAddressIndex(path);
+                    maxIndexHdPath = path;
+                }
+            }
+        }
+        AddressEntity address = mRepository.loadAddressBypath(maxIndexHdPath);
+        if (address == null) {
+            addAddress(maxIndexHdPath);
+        }
+    }
+
+    private void addAddress(String hdPath) {
+        String accountHdPath;
+        int pathIndex;
+        try {
+            AddressIndex index = CoinPath.parsePath(hdPath);
+            pathIndex = index.getValue();
+            accountHdPath = index.getParent().getParent().toString();
+        } catch (InvalidPathException e) {
+            e.printStackTrace();
+            return;
+        }
+
+        CoinEntity coin = mRepository.loadCoinSync(Coins.coinIdFromCoinCode(coinCode));
+        AccountEntity accountEntity = getAccountEntityByPath(accountHdPath, coin);
+        if (accountEntity == null) return;
+
+        List<AddressEntity> addressEntities = mRepository.loadAddressSync(coin.getCoinId());
+        Optional<AddressEntity> addressEntityOptional = addressEntities
+                .stream()
+                .filter(addressEntity -> addressEntity.getPath()
+                        .startsWith(accountEntity.getHdPath() + "/" + 0))
+                .max((o1, o2) -> o1.getPath().compareTo(o2.getPath()));
+
+        int index = -1;
+        if (addressEntityOptional.isPresent()) {
+            try {
+                AddressIndex addressIndex = CoinPath.parsePath(addressEntityOptional.get().getPath());
+                index = addressIndex.getValue();
+            } catch (InvalidPathException e) {
+                e.printStackTrace();
+            }
+        }
+
+        if (index < pathIndex) {
+            final CountDownLatch mLatch = new CountDownLatch(1);
+            addingAddress.postValue(true);
+            new AddAddressViewModel.AddAddressTask(coin, mRepository, mLatch::countDown,
+                    accountEntity.getExPub(), 0).execute(pathIndex - index);
+            try {
+                mLatch.await();
+                addingAddress.postValue(false);
+            } catch (InterruptedException e) {
+                e.printStackTrace();
+            }
+        }
+
+    }
+
+    public MutableLiveData<TxEntity> getObservableTx() {
+        return observableTx;
+    }
+}
