+++ coinlib/src/main/java/com/keystone/coinlib/Util.java
@@ -18,7 +18,6 @@
 package com.keystone.coinlib;
 
 import android.text.TextUtils;
-import android.util.Log;
 
 import com.keystone.coinlib.exception.InvalidPathException;
 import com.keystone.coinlib.path.AddressIndex;
@@ -42,7 +41,6 @@ import org.bouncycastle.util.Properties;
 import org.bouncycastle.util.encoders.Hex;
 
 import java.io.IOException;
-import java.nio.ByteBuffer;
 import java.nio.charset.StandardCharsets;
 import java.security.spec.ECPoint;
 import java.util.Arrays;
@@ -94,7 +92,7 @@ public class Util {
         if (address.toLowerCase().startsWith("bc")) {
             Bech32.Bech32Data decode = Bech32.decode(address);
             return Bech32.encode("tb", decode.data);
-        } else if (address.toLowerCase().startsWith("tb")){
+        } else if (address.toLowerCase().startsWith("tb")) {
             return address;
         }
         byte[] versionAndDataBytes = Base58.decodeChecked(address);
